Fix issue #9: setTime and submission time window issue
After calling setCurrentTime, we need to wait for actual time (an atomic reference in the same class, SandboxTimeProvider) to update itself via the time observer. Before, this logic was not there which caused in some cases that an exercise command following a setCurrentTime operation could potentially fail (submission time violated the corresponding time rules, it was "out if time window").

This commit solves the issue by waiting for an actualized "actual time" value (with Guava Monitor).